requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
ciso8601==2.2.0
Deprecated==1.2.13
idna==3.3
influxdb==5.3.1
influxdb-client==1.31.0
msgpack==1.0.4
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
Rx==3.2.0
six==1.16.0
urllib3==1.26.11
wrapt==1.14.1
